+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Changes in version 1.1.2
 
-- Add `insertCompound` function.
+- Add `insertCompound` and `deleteCompound` functions to add or remove compounds
+  from a `CompDb` or `IonDb`.
 
 ## Changes in version 1.1.1